refactorizacion en codigo y separacion de este mismo en archivos por separad:: aclaro que aun falta hacer ciertas mejora y el codigo puede ser mas escalable
